+++ offer-manager-domain/src/main/groovy/life/qbic/business/products/Converter.groovy
@@ -4,15 +4,9 @@ import life.qbic.business.logging.Logger
 import life.qbic.business.logging.Logging
 import life.qbic.datamodel.dtos.business.ProductCategory
 import life.qbic.datamodel.dtos.business.facilities.Facility
-import life.qbic.datamodel.dtos.business.services.DataStorage
-import life.qbic.datamodel.dtos.business.services.MetabolomicAnalysis
-import life.qbic.datamodel.dtos.business.services.PrimaryAnalysis
+import life.qbic.datamodel.dtos.business.services.*
 import life.qbic.datamodel.dtos.business.services.Product
-import life.qbic.datamodel.dtos.business.services.ProductUnit
-import life.qbic.datamodel.dtos.business.services.ProjectManagement
-import life.qbic.datamodel.dtos.business.services.ProteomicAnalysis
-import life.qbic.datamodel.dtos.business.services.SecondaryAnalysis
-import life.qbic.datamodel.dtos.business.services.Sequencing
+
 
 /**
  * <h1>Converter for {@link life.qbic.datamodel.dtos.business.services.Product}</h1>
@@ -22,7 +16,7 @@ import life.qbic.datamodel.dtos.business.services.Sequencing
  *
  * @since 1.0.0
  *
-*/
+ */
 class Converter {
 
     private static final Logging log = Logger.getLogger(this.class)
@@ -36,8 +30,9 @@ class Converter {
      * @param unitPrice The unit price of the product
      * @param unit The unit in which the product is measured
      * @return a product
-     * @deprecated please us {@link life.qbic.business.products.Converter#createProduct(life.qbic.datamodel.dtos.business.ProductCategory, java.lang.String, java.lang.String, double, life.qbic.datamodel.dtos.business.services.ProductUnit, life.qbic.datamodel.dtos.business.facilities.Facility)} instead
+     * @deprecated please use {@link life.qbic.business.products.Converter#createProduct(life.qbic.datamodel.dtos.business.ProductCategory, java.lang.String, java.lang.String, double, double, life.qbic.datamodel.dtos.business.services.ProductUnit, life.qbic.datamodel.dtos.business.facilities.Facility)} instead
      */
+    @Deprecated
     static Product createProduct(ProductCategory category, String name, String description, double unitPrice, ProductUnit unit){
         long runningNumber = 0 //todo it should be possible to create products without a running number
         return createProductWithVersion(category,name,description,unitPrice,unit,runningNumber)
@@ -49,15 +44,16 @@ class Converter {
      * @param category The products category which determines what kind of products is created
      * @param description The description of the product
      * @param name The name of the product
-     * @param unitPrice The unit price of the product
+     * @param internalUnitPrice The unit price for internal customers
+     * @param externalUnitPrice The unit price for external customers
      * @param unit The unit in which the product is measured
-     * @param facility The facility providing the product
+     * @param serviceProvider The serviceProvider providing the product
      * @return a new product
      * @since 1.1.0
      */
-    static Product createProduct(ProductCategory category, String name, String description, double unitPrice, ProductUnit unit, Facility facility){
+    static Product createProduct(ProductCategory category, String name, String description, double internalUnitPrice, double externalUnitPrice, ProductUnit unit, Facility serviceProvider){
         long runningNumber = 0 //todo it should be possible to create products without a running number
-        return createProductWithVersion(category,name,description,unitPrice,unit,runningNumber, facility)
+        return createProductWithVersion(category, name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
     }
 
     /**
@@ -70,8 +66,9 @@ class Converter {
      * @param unit The unit in which the product is measured
      * @param runningNumber The running version number of the product
      * @return a product
-     * @deprecated please use {@link life.qbic.business.products.Converter#createProductWithVersion(life.qbic.datamodel.dtos.business.ProductCategory, java.lang.String, java.lang.String, double, life.qbic.datamodel.dtos.business.services.ProductUnit, long, life.qbic.datamodel.dtos.business.facilities.Facility) } instead
+     * @deprecated please use {@link life.qbic.business.products.Converter#createProductWithVersion(life.qbic.datamodel.dtos.business.ProductCategory, java.lang.String, java.lang.String, double, double, life.qbic.datamodel.dtos.business.services.ProductUnit, long, life.qbic.datamodel.dtos.business.facilities.Facility) } instead
      */
+    @Deprecated
     static Product createProductWithVersion(ProductCategory category, String name, String description, double unitPrice, ProductUnit unit, long runningNumber){
         Product product = null
         switch (category) {
@@ -109,36 +106,36 @@ class Converter {
      * @param category The products category which determines what kind of products is created
      * @param description The description of the product
      * @param name The name of the product
-     * @param unitPrice The unit price of the product
+     * @param internalUnitPrice The unit price of the product
      * @param unit The unit in which the product is measured
      * @param runningNumber The running version number of the product
-     * @param facility The facility providing the product
+     * @param serviceProvider The serviceProvider providing the product
      * @return a product
      * @since 1.1.0
      */
-    static Product createProductWithVersion(ProductCategory category, String name, String description, double unitPrice, ProductUnit unit, long runningNumber, Facility facility){
+    static Product createProductWithVersion(ProductCategory category, String name, String description, double internalUnitPrice, double externalUnitPrice, ProductUnit unit, long runningNumber, Facility serviceProvider){
         Product product = null
         switch (category) {
             case ProductCategory.DATA_STORAGE:
-                product = new DataStorage(name, description, 0.0, 0.0, unit, runningNumber, facility)
+                product = new DataStorage(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.PRIMARY_BIOINFO:
-                product = new PrimaryAnalysis(name, description, unitPrice,unit, runningNumber.toString())
+                product = new PrimaryAnalysis(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.PROJECT_MANAGEMENT:
-                product = new ProjectManagement(name, description, unitPrice,unit, runningNumber.toString())
+                product = new ProjectManagement(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.SECONDARY_BIOINFO:
-                product = new SecondaryAnalysis(name, description, unitPrice,unit, runningNumber.toString())
+                product = new SecondaryAnalysis(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.SEQUENCING:
-                product = new Sequencing(name, description, unitPrice,unit, runningNumber.toString())
+                product = new Sequencing(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.PROTEOMIC:
-                product = new ProteomicAnalysis(name, description, unitPrice,unit, runningNumber.toString())
+                product = new ProteomicAnalysis(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             case ProductCategory.METABOLOMIC:
-                product = new MetabolomicAnalysis(name, description, unitPrice,unit, runningNumber.toString())
+                product = new MetabolomicAnalysis(name, description, internalUnitPrice, externalUnitPrice, unit, runningNumber, serviceProvider)
                 break
             default:
                 log.warn("Unknown product category $category")
@@ -166,16 +163,42 @@ class Converter {
 
     static life.qbic.business.products.Product convertDTOtoProduct(Product product){
         ProductCategory category = getCategory(product)
-        return new life.qbic.business.products.Product.Builder(category,
-                                                                product.productName,
-                                                                product.description,
-                                                                product.unitPrice,
-                                                                product.unit)
-                                                                .build()
+        if (! usesInternalAndExternalPricing(product)) { // we used an old constructor
+            return new life.qbic.business.products.Product.Builder(category,
+                    product.productName,
+                    product.description,
+                    product.unitPrice,
+                    product.unit)
+                    .build()
+        } else {
+            return new life.qbic.business.products.Product.Builder(category,
+                    product.productName,
+                    product.description,
+                    product.internalUnitPrice,
+                    product.externalUnitPrice,
+                    product.unit,
+                    product.serviceProvider)
+                    .build()
+        }
 
     }
 
     static Product convertProductToDTO(life.qbic.business.products.Product product){
-        return createProductWithVersion(product.category,product.name, product.description, product.unitPrice, product.unit, product.id.uniqueId)
+        Product result
+
+        if (! usesInternalAndExternalPricing(product)) { // we used an old constructor
+            result = createProductWithVersion(product.category,product.name, product.description, product.unitPrice, product.unit, product.id.uniqueId)
+        } else { // we used the new constructor
+            result = createProductWithVersion(product.category,product.name, product.description, product.internalUnitPrice, product.externalUnitPrice, product.unit, product.id.uniqueId, product.serviceProvider)
+        }
+        return result
+    }
+
+    private static boolean usesInternalAndExternalPricing(life.qbic.business.products.Product product) {
+        return ((product.internalUnitPrice != 0 || product.externalUnitPrice != 0) && product.unitPrice == 0)
+    }
+
+    private static boolean usesInternalAndExternalPricing(Product product) {
+        return ((product.internalUnitPrice != 0 || product.externalUnitPrice != 0) && product.unitPrice == 0)
     }
 }